+++ src/main/java/org/agenda/event/repository/EventRepository.java
@@ -7,10 +7,10 @@ import java.util.Optional;
 
 public interface EventRepository {
     CalendarEvent save(CalendarEvent event);
-    Optional<CalendarEvent> findById(int id);
+    Optional<CalendarEvent> findById(long id);
     List<CalendarEvent> findAll();
     List<CalendarEvent> findUpcomingEvents(int intervalDays);
 
     boolean updateById(CalendarEvent event);
-    boolean deleteById(int id);
+    boolean deleteById(long id);
 }
